requirements.txt
absl-py==2.1.0
afterglowpy==0.7.3
antiprism_python==0.1.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
astroplan==0.9.1
astropy==6.0.0
astropy-healpix==1.0.2
astropy-iers-data==0.2024.3.4.0.30.17
astroquery==0.4.7
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blosc2==2.5.1
cdshealpix==0.6.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.0
cplex==22.1.1.0
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.9
dask==2024.2.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docplex==2.27.239
dorado-scheduling @ git+https://github.com/nasa/dorado-scheduling@6e70c8eb08751ad0e08d8b452a3c4bf1869b3554
dust_extinction==1.3
dustmaps==1.0.13
ephem==4.1.5
exceptiongroup==1.2.0
executing==2.0.1
extinction==0.4.6
fasteners==0.19
fastjsonschema==2.20.0
fonttools==4.49.0
fqdn==1.5.1
fsspec==2024.2.0
future==1.0.0
george==0.4.1
gurobipy==11.0.2
gwemlightcurves==0.0.5
gwemopt==0.3.4
h11==0.14.0
h5py==3.10.0
healpy==1.16.6
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
igwn-auth-utils==1.1.0
immutabledict==4.2.0
importlib_metadata==7.0.2
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.3
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keyring==25.3.0
kiwisolver==1.4.5
lalsuite==7.21
ligo-gracedb==2.12.0
ligo-segments==1.4.0
ligo.skymap==1.1.2
locket==1.0.0
looseversion==1.3.0
lscsoft-glue==4.0.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mistune==3.0.2
mocpy==0.16.2
more-itertools==10.4.0
msgpack==1.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.1
notebook_shim==0.2.4
numcodecs==0.12.1
numexpr==2.10.1
numpy==1.26.4
ortools==9.10.4067
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.1
penquins==2.4.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.4.0
progressbar2==4.4.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptemcee==1.0.0
ptyprocess==0.7.0
PuLP==2.9.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
PyAstronomy==0.22.0
pycparser==2.21
pyerfa==2.0.1.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.6.3
pymultinest==2.12
pyOpenSSL==24.0.0
pyparsing==3.1.2
pytest==8.0.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-ligo-lw==1.8.3
python-utils==3.8.2
pytz==2024.1
pyvo==1.5.2
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
quantities==0.15.0
radbelt==0.1.7
referencing==0.35.1
regions==0.8
reproject==0.13.0
requests==2.32.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
safe-netrc==1.0.1
scikit-learn==1.5.0
scipy==1.12.0
scitokens==1.8.1
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
sfdmap==0.1.1
sgp4==2.23
shapely==2.0.3
six==1.16.0
sncosmo==2.10.4
sniffio==1.3.1
soupsieve==2.5
spiceypy==6.0.0
stack-data==0.6.3
synphot==1.3.post0
tables==3.9.2
terminado==0.18.1
threadpoolctl==3.3.0
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.3
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
VOEventLib==1.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
zarr==2.17.1
zipp==3.17.0

